move react code to presentation layer
