+++ apps/web/src/presentation/components/launchpad/AugmentedBondingCurveSettings.tsx
