+++ apps/web/src/presentation/components/launchpad/VotingSettings.tsx
