+++ apps/web/src/presentation/components/shared/Footer.tsx
