+++ apps/web/src/presentation/components/shared/footer/SubscribeInput.css
