+++ apps/web/src/presentation/theme/components/slider.ts
